+++ HISTORY.md
@@ -1,5 +1,6 @@
 # jacman-core
 ## [v1.0]
+* extracted from JacMan
 
 ## [v1.1]
 * 'Force gi' button